requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
google-generativeai
redis
requests